+++ crates/user_input/src/read_line.rs
@@ -1,7 +1,6 @@
-use std::io::Write;
-
 use eyre::Context;
 use eyre::Result;
+use std::io::Write;
 use tokio::io;
 use tokio::io::AsyncBufReadExt;
 use tokio::io::BufReader;