+++ crates/entrypoint/src/interactive/browse_users.rs
@@ -1,7 +1,7 @@
 use cloud_terrastodon_azure::prelude::fetch_all_users;
-use cloud_terrastodon_user_input::prelude::Choice;
-use cloud_terrastodon_user_input::prelude::FzfArgs;
-use cloud_terrastodon_user_input::prelude::pick_many;
+use cloud_terrastodon_user_input::Choice;
+use cloud_terrastodon_user_input::FzfArgs;
+use cloud_terrastodon_user_input::pick_many;
 use eyre::Result;
 use itertools::Itertools;
 use tracing::info;