+++ crates/user_input/examples/are_you_sure.rs
@@ -1,4 +1,4 @@
-use cloud_terrastodon_user_input::prelude::are_you_sure;
+use cloud_terrastodon_user_input::are_you_sure;
 
 fn main() -> eyre::Result<()> {
     if are_you_sure("This will delete the universe. Are you sure you want to proceed?".to_string())? {