+++ crates/user_input/examples/multiline.rs
@@ -1,4 +1,4 @@
-use cloud_terrastodon_user_input::prelude::pick_many;
+use cloud_terrastodon_user_input::pick_many;
 
 pub fn main() -> eyre::Result<()> {
     let choices = vec![