+++ crates/user_input/src/are_you_sure.rs
@@ -1,5 +1,5 @@
-use crate::prelude::FzfArgs;
-use crate::prelude::pick;
+use crate::FzfArgs;
+use crate::pick;
 
 pub fn are_you_sure(message: impl AsRef<str>) -> eyre::Result<bool> {
     let choices = ["No", "Yes"];