+++ crates/user_input/src/lib.rs
@@ -2,8 +2,6 @@ mod are_you_sure;
 mod fzf;
 mod read_line;
 
-pub mod prelude {
-    pub use crate::are_you_sure::*;
-    pub use crate::fzf::*;
-    pub use crate::read_line::*;
-}
+pub use crate::are_you_sure::*;
+pub use crate::fzf::*;
+pub use crate::read_line::*;